refator: refactoring the tests with the new clients
