+++ src/utils/__tests__/api/helpers/ecommerce/ecommercePaymentRequestsHelper.test.ts
@@ -1,6 +1,6 @@
 import { pipe } from "fp-ts/function";
 import * as E from "fp-ts/Either";
-import { RptId } from "../../../../../../generated/definitions/payment-ecommerce-v3/RptId";
+import { RptId } from "../../../../../../generated/definitions/payment-ecommerce-auth-v1/RptId";
 import {
   getCartResponseMock,
   getPaymentRequestInfoResponseMock,
@@ -8,7 +8,7 @@ import {
 } from "../../../../../utils/testing/testUtils";
 import {
   apiPaymentEcommerceClient,
-  apiPaymentEcommerceClientV3,
+  apiPaymentEcommerceAuthClientV1,
 } from "../../../../api/client";
 import { getCarts, getEcommercePaymentInfoTask } from "../../../../api/helper";
 import { ErrorsType } from "../../../../errors/checkErrorsModel";
@@ -31,7 +31,7 @@ jest.mock("../../../../api/client", () => ({
     GetCarts: jest.fn(),
     getPaymentRequestInfo: jest.fn(),
   },
-  apiPaymentEcommerceClientV3: {
+  apiPaymentEcommerceAuthClientV1: {
     getPaymentRequestInfoV3: jest.fn(),
   },
 }));
@@ -111,7 +111,7 @@ describe("Ecommerce payment requests helper - getEcommercePaymentInfoTask tests"
   it("Should return paymentInfo when api return correct value on v3 api", async () => {
     (getSessionItem as jest.Mock).mockReturnValue("authToken");
     (
-      apiPaymentEcommerceClientV3.getPaymentRequestInfoV3 as jest.Mock
+      apiPaymentEcommerceAuthClientV1.getPaymentRequestInfoAuth as jest.Mock
     ).mockReturnValue(
       Promise.resolve({
         right: {
@@ -132,7 +132,7 @@ describe("Ecommerce payment requests helper - getEcommercePaymentInfoTask tests"
   it("Should return error when api fail on v3 api", async () => {
     (getSessionItem as jest.Mock).mockReturnValue("authToken");
     (
-      apiPaymentEcommerceClientV3.getPaymentRequestInfoV3 as jest.Mock
+      apiPaymentEcommerceAuthClientV1.getPaymentRequestInfoAuth as jest.Mock
     ).mockRejectedValue("Api error");
     const result = await pipe(
       getEcommercePaymentInfoTask(
@@ -150,7 +150,7 @@ describe("Ecommerce payment requests helper - getEcommercePaymentInfoTask tests"
   it("Should return ErrorsType.GENERIC_ERROR when api return 5xx on v3 api", async () => {
     (getSessionItem as jest.Mock).mockReturnValue("authToken");
     (
-      apiPaymentEcommerceClientV3.getPaymentRequestInfoV3 as jest.Mock
+      apiPaymentEcommerceAuthClientV1.getPaymentRequestInfoAuth as jest.Mock
     ).mockReturnValue(
       Promise.resolve({
         right: {
@@ -178,7 +178,7 @@ describe("Ecommerce payment requests helper - getEcommercePaymentInfoTask tests"
   it("Should return SESSION_EXPIRED when api return 401 on v3 api", async () => {
     (getSessionItem as jest.Mock).mockReturnValue("authToken");
     (
-      apiPaymentEcommerceClientV3.getPaymentRequestInfoV3 as jest.Mock
+      apiPaymentEcommerceAuthClientV1.getPaymentRequestInfoAuth as jest.Mock
     ).mockReturnValue(
       Promise.resolve({
         right: {
